Perf: Function for date and required fields validation
